refactor: split OpenRouter evaluation support
